Pass the value type around for decoding
I started off doing this, decided not to and now I realise I should
again beacuse it's the only sane way to get dates around.

This does remove the hack I'd put in place for signed / unsigned binary
that I hadn't figured out how it would actually be implemented for the
general parser so that's good?