Add packet parsing
First steps to actually being able to do anything useful with all this
code which is nice.

I'm not sure this is a good final form for this system but I'm just
going to throw things at the wall until it works and then rebuild it
into something sensible which is a fine development method.

I'm also really not sure what the other packet types libmbus supports
are since they don't seem to be showing up in my specs