Convert DIB parsing
I dunno if it works or not since I don't have any tests, but it should.

I decided not to continue the silliness around tariff/device being
options since it's completely useless.

Presumably I'm going to need to reverse engineer it for the XML export,
but that shouldn't be too hard. If both of them are zero and the storage
is <2 then they should be null.