Implement all the remaining VIF parsing
Comes with a ton of confusion because 0xFB is the "first extension" but
actually refers to Table 14 "Alternate extended VIF-code table" while
0xFD is the "second extension" but refers to Table 12 "Main VIFE-code
extension table" which sounds like it should actually be the first
extension. Lots of debugging and anger on that one.

However, I ended up fixing that, adding extra debug info and updating
the macro to also support `p` in place of `n`. Fun fun fun.

Only 22 tests failing from the test suite now!