"Support" the "Signature" field
EN 13757-7:2018 7.5.8 refers to this as the Configuration Field (CF) and
has a lot to say about the various encryption protocols it indicates via
a bitfield and what to do with those bits and how they will affect all
subsequent parsing of the frame, including indicating extra headers.

https://m-bus.com/documentation-wired/06-application-layer refers to
this as the "Signature" field and states "The Signature remains reserved
for future encryptation applications, and until then is allocated the
value 00 00 h."

libmbus refers to this as the "Signature" field and treats it as a
random hexadecimal string put in as decoration.

Thankfully their test data (accidentally?) lines up with the reserved
encryption protocols and so I can maintain compatability with them by
once again adding a "Reserved" option type that should have been an
error and stuffing the whole 16 bits in there.