Parse the application message data
The spec is extremely unclear about what the meter is supposed to
respond with. The example has it echoing the config back, which seems
like it would be completely useless.

I'm going to have to find a meter that supports this and muck about with
it until I understand what it's doing.