De-Modal the primary parsers
As far as I can tell, the direction Winnow is going in expects you to
not actually need `ErrMode` unless your parser deals with partial data,
or needs to be able to cut backtracking.

Since the MBus protocol doesn't do the kind of backtrackable prefixes
that this would matter for, I reckon we can simplify matters and stay
ahead of the wave by doing this.