Track assertion text in non-debug builds
tbh I should probably make this into a pull request for Winnow since it
seems silly that `ContextError`s don't track this.

It's also mildly silly how much effort I had to go through to make this
work, with the custom context type and formatter and so on, but I
suppose it's useful for other things too.